+++ NINA.Plugin.Assistant/NINA.Plugin.Assistant/Sequencer/TargetSchedulerContainer.cs
@@ -406,7 +406,6 @@ namespace Assistant.NINAPlugin.Sequencer {
         }
 
         public override bool Validate() {
-            TSLogger.Debug("TargetSchedulerContainer: Validate");
             var issues = new List<string>();
 
             var triggers = GetTriggersSnapshot();